refactor: simplify History page management
