+++ src/renderer/src/App.tsx
@@ -6,7 +6,6 @@ import './App.css';
 import { Editor } from './Editor/Editor';
 import { EditorIndex } from './Editor/index';
 import { NavBar } from './components/navigation/NavBar';
-import { HistoryIndex } from './pages/History';
 import { History } from './pages/History/History';
 import { Options } from './pages/Options/Options';
 import { ThemeContext, themes } from './personalization/theme';
@@ -27,7 +26,7 @@ function App() {
           <Route path="/" element={<EditorIndex />} />
           <Route path="/edit" element={<EditorIndex />} />
           <Route path="/edit/:documentId" element={<Editor />} />
-          <Route path="history" element={<HistoryIndex />}>
+          <Route path="history" element={<History />}>
             <Route path=":documentId" element={<History />} />
           </Route>
           <Route path="/options" element={<Options />} />
